+++ scripts/winesapos-setup.sh
@@ -139,7 +139,7 @@ if [ $? -eq 0 ]; then
     qdbus ${kdialog_dbus} /ProgressDialog org.kde.kdialog.ProgressDialog.close
 fi
 
-graphics_selected=$(kdialog --title "winesapOS First-Time Setup" --menu "Select your desired graphics driver..." amd AMD intel Intel nvidia NVIDIA virtualbox VirtualBox vmware VMware)
+graphics_selected=$(kdialog --title "winesapOS First-Time Setup" --menu "Select your desired graphics driver..." amd AMD intel Intel nvidia-new "NVIDIA (New, Maxwell and newer)" nvidia-old "NVIDIA (Old, Kepler and newer)" virtualbox VirtualBox vmware VMware)
 # Keep track of the selected graphics drivers for upgrade purposes.
 echo ${graphics_selected} | sudo tee /etc/winesapos/graphics
 kdialog_dbus=$(kdialog --title "winesapOS First-Time Setup" --progressbar "Please wait for the graphics driver to be installed..." 2 | cut -d" " -f1)
@@ -151,13 +151,29 @@ elif [[ "${graphics_selected}" == "intel" ]]; then
     sudo pacman -S --noconfirm \
       extra/intel-media-driver \
       extra/intel-compute-runtime
-elif [[ "${graphics_selected}" == "nvidia" ]]; then
+elif [[ "${graphics_selected}" == "nvidia-new" ]]; then
     sudo pacman -S --noconfirm \
       extra/nvidia-dkms \
       extra/nvidia-utils \
       multilib/lib32-nvidia-utils \
       extra/opencl-nvidia \
       multilib/lib32-opencl-nvidia
+
+    # Block the loading of conflicting open source NVIDIA drivers.
+    sudo touch /etc/modprobe.d/winesapos-nvidia.conf
+    echo "blacklist nouveau
+blacklist nvidiafb
+blacklist nv
+blacklist rivafb
+blacklist rivatv
+blacklist uvcvideo" | sudo tee /etc/modprobe.d/winesapos-nvidia.conf
+elif [[ "${graphics_selected}" == "nvidia-old" ]]; then
+    ${CMD_YAY_INSTALL} \
+      nvidia-470xx-dkms \
+      nvidia-470xx-utils \
+      lib32-nvidia-470xx-utils \
+      opencl-nvidia-470xx \
+      lib32-opencl-nvidia-470xx
 
     # Block the loading of conflicting open source NVIDIA drivers.
     sudo touch /etc/modprobe.d/winesapos-nvidia.conf