+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Add support for Framework laptops.
 - Change the fast Pacman mirror service to run as part of the first-time setup instead of the installation.
 - Add support for Microsoft Surface laptops.
+- Add support for older NVIDIA Kepler cards.
 
 ## 3.2.1 - The SteamOS 3.4 Update